refactor: account switcher menu item into its own component
