+++ packages/shared/components/molecules/index.js
@@ -1,3 +1,5 @@
 export { default as TransactionDetails } from './TransactionDetails.svelte'
 export { default as ExpirationTimePicker } from './ExpirationTimePicker.svelte'
 export { default as AssetList } from './AssetList.svelte'
+
+export * from './menu-items'
