+++ packages/shared/components/molecules/menu-items/index.js
@@ -1,0 +1,1 @@
+export { default as AccountSwitcherMenuItem } from './AccountSwitcherMenuItem.svelte'
